updated posts_tags migration to check for the table if it exists

+++ db/migrate/20131210204218_alter_posts_tags_table_no_timestamps.rb
@@ -1,6 +1,6 @@
 class AlterPostsTagsTableNoTimestamps < ActiveRecord::Migration
   def up
-    drop_table :posts_tags
+    drop_table :posts_tags if ActiveRecord::Base.connection.table_exists? :posts_tags
     create_table :posts_tags, :id => false do |t|
       t.integer :post_id, :null => false
       t.integer :tag_id,  :null => false
